requirements.txt
aniso8601==7.0.0
asgiref==3.2.5
Django==2.1.4
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.1.5
django-prometheus==2.0.0
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.1
graphql-relay==2.0.1
prometheus-client==0.7.1
promise==2.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
